r/aws_fsx_data_repository_association: Move functions around.
