+++ internal/service/fsx/find.go
@@ -92,75 +92,6 @@ func findFileCacheByID(ctx context.Context, conn *fsx.FSx, id string) (*fsx.File
 	return fileCaches[0], nil
 }
 
-func findDataRepositoryAssociationsByIDs(ctx context.Context, conn *fsx.FSx, ids []*string) ([]*fsx.DataRepositoryAssociation, error) {
-	input := &fsx.DescribeDataRepositoryAssociationsInput{
-		AssociationIds: ids,
-	}
-	var dataRepositoryAssociations []*fsx.DataRepositoryAssociation
-
-	err := conn.DescribeDataRepositoryAssociationsPagesWithContext(ctx, input, func(page *fsx.DescribeDataRepositoryAssociationsOutput, lastPage bool) bool {
-		if page == nil {
-			return !lastPage
-		}
-		dataRepositoryAssociations = append(dataRepositoryAssociations, page.Associations...)
-		return !lastPage
-	})
-
-	if tfawserr.ErrCodeEquals(err, fsx.ErrCodeDataRepositoryAssociationNotFound) {
-		return nil, &retry.NotFoundError{
-			LastError:   err,
-			LastRequest: input,
-		}
-	}
-	if err != nil {
-		return nil, err
-	}
-	if len(dataRepositoryAssociations) == 0 || dataRepositoryAssociations[0] == nil {
-		return nil, tfresource.NewEmptyResultError(input)
-	}
-
-	return dataRepositoryAssociations, nil
-}
-
-func FindDataRepositoryAssociationByID(ctx context.Context, conn *fsx.FSx, id string) (*fsx.DataRepositoryAssociation, error) {
-	input := &fsx.DescribeDataRepositoryAssociationsInput{
-		AssociationIds: []*string{aws.String(id)},
-	}
-
-	var associations []*fsx.DataRepositoryAssociation
-
-	err := conn.DescribeDataRepositoryAssociationsPagesWithContext(ctx, input, func(page *fsx.DescribeDataRepositoryAssociationsOutput, lastPage bool) bool {
-		if page == nil {
-			return !lastPage
-		}
-
-		associations = append(associations, page.Associations...)
-
-		return !lastPage
-	})
-
-	if tfawserr.ErrCodeEquals(err, fsx.ErrCodeDataRepositoryAssociationNotFound) {
-		return nil, &retry.NotFoundError{
-			LastError:   err,
-			LastRequest: input,
-		}
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	if len(associations) == 0 || associations[0] == nil {
-		return nil, tfresource.NewEmptyResultError(input)
-	}
-
-	if count := len(associations); count > 1 {
-		return nil, tfresource.NewTooManyResultsError(count, input)
-	}
-
-	return associations[0], nil
-}
-
 func FindStorageVirtualMachineByID(ctx context.Context, conn *fsx.FSx, id string) (*fsx.StorageVirtualMachine, error) {
 	input := &fsx.DescribeStorageVirtualMachinesInput{
 		StorageVirtualMachineIds: []*string{aws.String(id)},
