+++ internal/service/fsx/wait.go
@@ -203,72 +203,6 @@ func waitFileSystemDeleted(ctx context.Context, conn *fsx.FSx, id string, timeou
 	return nil, err
 }
 
-func waitDataRepositoryAssociationCreated(ctx context.Context, conn *fsx.FSx, id string, timeout time.Duration) (*fsx.DataRepositoryAssociation, error) {
-	stateConf := &retry.StateChangeConf{
-		Pending: []string{fsx.DataRepositoryLifecycleCreating},
-		Target:  []string{fsx.DataRepositoryLifecycleAvailable},
-		Refresh: statusDataRepositoryAssociation(ctx, conn, id),
-		Timeout: timeout,
-		Delay:   30 * time.Second,
-	}
-
-	outputRaw, err := stateConf.WaitForStateContext(ctx)
-
-	if output, ok := outputRaw.(*fsx.DataRepositoryAssociation); ok {
-		if status, details := aws.StringValue(output.Lifecycle), output.FailureDetails; status == fsx.DataRepositoryLifecycleFailed && details != nil {
-			tfresource.SetLastError(err, errors.New(aws.StringValue(output.FailureDetails.Message)))
-		}
-
-		return output, err
-	}
-
-	return nil, err
-}
-
-func waitDataRepositoryAssociationUpdated(ctx context.Context, conn *fsx.FSx, id string, timeout time.Duration) (*fsx.DataRepositoryAssociation, error) {
-	stateConf := &retry.StateChangeConf{
-		Pending: []string{fsx.DataRepositoryLifecycleUpdating},
-		Target:  []string{fsx.DataRepositoryLifecycleAvailable},
-		Refresh: statusDataRepositoryAssociation(ctx, conn, id),
-		Timeout: timeout,
-		Delay:   30 * time.Second,
-	}
-
-	outputRaw, err := stateConf.WaitForStateContext(ctx)
-
-	if output, ok := outputRaw.(*fsx.DataRepositoryAssociation); ok {
-		if status, details := aws.StringValue(output.Lifecycle), output.FailureDetails; status == fsx.DataRepositoryLifecycleFailed && details != nil {
-			tfresource.SetLastError(err, errors.New(aws.StringValue(output.FailureDetails.Message)))
-		}
-
-		return output, err
-	}
-
-	return nil, err
-}
-
-func waitDataRepositoryAssociationDeleted(ctx context.Context, conn *fsx.FSx, id string, timeout time.Duration) (*fsx.DataRepositoryAssociation, error) {
-	stateConf := &retry.StateChangeConf{
-		Pending: []string{fsx.DataRepositoryLifecycleAvailable, fsx.DataRepositoryLifecycleDeleting},
-		Target:  []string{},
-		Refresh: statusDataRepositoryAssociation(ctx, conn, id),
-		Timeout: timeout,
-		Delay:   30 * time.Second,
-	}
-
-	outputRaw, err := stateConf.WaitForStateContext(ctx)
-
-	if output, ok := outputRaw.(*fsx.DataRepositoryAssociation); ok {
-		if status, details := aws.StringValue(output.Lifecycle), output.FailureDetails; status == fsx.DataRepositoryLifecycleFailed && details != nil {
-			tfresource.SetLastError(err, errors.New(aws.StringValue(output.FailureDetails.Message)))
-		}
-
-		return output, err
-	}
-
-	return nil, err
-}
-
 func waitStorageVirtualMachineCreated(ctx context.Context, conn *fsx.FSx, id string, timeout time.Duration) (*fsx.StorageVirtualMachine, error) {
 	stateConf := &retry.StateChangeConf{
 		Pending: []string{fsx.StorageVirtualMachineLifecycleCreating, fsx.StorageVirtualMachineLifecyclePending},
